feat: added more methods to gig

+++ src/main/java/com/allesys/findmeworks/service/LocationService/LocationServiceImpl.java
@@ -5,7 +5,9 @@ import java.util.UUID;
 
 import com.allesys.findmeworks.dto.LocationDTO;
 import com.allesys.findmeworks.entity.User;
+import com.allesys.findmeworks.repository.GigRepository;
 import com.allesys.findmeworks.repository.UserRepository;
+import com.allesys.findmeworks.service.UserService.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,10 +17,18 @@ import com.allesys.findmeworks.repository.LocationRepository;
 @Service
 public class LocationServiceImpl implements LocationService {
     private final LocationRepository locationRepository;
+    private final GigRepository gigRepository;
+    private final UserService userService;
 
     @Autowired
-    public LocationServiceImpl(LocationRepository locationRepository, UserRepository userRepository) {
+    public LocationServiceImpl(
+            LocationRepository locationRepository,
+            GigRepository gigRepository,
+            UserService userService
+    ) {
         this.locationRepository = locationRepository;
+        this.gigRepository = gigRepository;
+        this.userService = userService;
     }
 
     @Override
@@ -28,6 +38,7 @@ public class LocationServiceImpl implements LocationService {
         location.setLocationName(locationRequest.getLocationName());
         location.setLatitude(locationRequest.getLatitude());
         location.setLongitude(locationRequest.getLongitude());
+
         return this.locationRepository.save(location);
     }
 
@@ -43,7 +54,16 @@ public class LocationServiceImpl implements LocationService {
 
     @Override
     public void deleteLocation(UUID id) {
-        this.locationRepository.deleteById(id);
+        Location location = this.locationRepository.findById(id).orElse(null);
+
+        if (location != null) {
+            UUID userId = location.getUserId();
+
+            userService.removeLocationFromUser(userId, id);
+
+
+            this.locationRepository.deleteById(id);
+        }
     }
 
 }
